requirements.txt
altgraph==0.17.4
attrs==24.2.0
auto-py-to-exe==2.44.1
blinker==1.8.2
bottle==0.13.1
bottle-websocket==0.2.9
branca==0.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
distlib==0.3.8
Eel==0.17.0
filelock==3.16.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
folium==0.17.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.1.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.1
outcome==1.3.0.post0
packaging==24.1
pefile==2024.8.26
platformdirs==4.3.6
pycparser==2.22
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.4
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.32.3
selenium==4.25.0
setuptools==75.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.5
waitress==3.0.0
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
whichcraft==0.6.1
wsproto==1.2.0
xyzservices==2024.9.0
zope.event==5.0
zope.interface==7.0.3
